requirements.txt
# Core dependencies
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.21.0,<2.0.0
scipy>=1.7.0
scikit-image>=0.19.0
SimpleITK>=2.2.0
tqdm>=4.64.0
pandas>=1.4.0
matplotlib>=3.5.0

# nnU-Net and nnInteractive dependencies
nnunetv2>=2.2
batchgenerators>=0.25
acvl-utils>=0.2

# Medical imaging
monai>=1.3.0
nibabel>=3.2.0
antspyx>=0.3.0

# Interaction and visualization
opencv-python>=4.5.0
Pillow>=9.0.0

# Configuration and I/O
PyYAML>=6.0
h5py>=3.7.0
imageio>=2.19.0

# Hugging Face for model weights
huggingface-hub>=0.16.0

# Optional but recommended for performance
tensorboard>=2.10.0
wandb>=0.13.0

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950